# lab5/swagger_done.yaml
openapi: 3.0.3
info:
  version: '1.0'
  title: 'Countries API Laravel & JWT'
  description: Lab006 PAB
servers:
  - description: Laravel
    url: http://localhost:8000/api
paths:
...
  /auth/forgot-password:
    post:
      tags:
        - auth
      summary: Requests a password reset token by providing the user's email
      description: |
        This endpoint allows a user to request a password reset token by providing their email. 
        If the email exists in the system, a token will be generated and logged (in place of sending an email).
        The user can then use this token to reset their password via another endpoint.
      operationId: forgotPassword
      requestBody:
        description: User email to receive the password reset token
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
      responses:
        '202':
          description: Token request accepted, token logged successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: null
        '404':
          description: User not found with the provided email address
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User not found
        '422':
          description: Invalid or missing email format
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Invalid email format or email is required
  /auth/reset-password:
    post:
      tags:
        - auth
      summary: Resets the user's password using the provided token and new password
      description: |
        This endpoint allows a user to reset their password by providing a valid reset token, their email address, and a new password. 
        The reset token should have been generated earlier (via the forgot password endpoint).
      operationId: resetPassword
      requestBody:
        description: Request body for resetting the user's password
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
                - email
                - password
              properties:
                token:
                  type: string
                  example: reset-token-here
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  minLength: 3
                  example: newpassword123
      responses:
        '200':
          description: Password reset successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Password has been successfully reset."
        '400':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Invalid or expired token"
        '422':
          description: Invalid or missing input data (e.g., invalid email or password)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Invalid email or password format."
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []          